requirements.txt
allure-pytest==2.13.5
allure-python-commons==2.13.5
attrs==25.1.0
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
colorama==0.4.6
h11==0.14.0
idna==3.10
iniconfig==2.0.0
packaging==24.2
pluggy==1.5.0
pycparser==2.22
pytest==8.3.4
pytest-repeat==0.9.3
python-dotenv==1.0.1
selenium==4.29.0
typing_extensions==4.12.2
urllib3==2.3.0
webdriver-manager==4.0.2
